Get rid of flocker.node.gear.
