+++ flocker/node/functional/test_docker.py
@@ -28,8 +28,7 @@ from ...testtools import (
     random_name)
 
 from ..test.test_docker import make_idockerclient_tests
-from .._docker import DockerClient
-from ..gear import PortMap, GearEnvironment
+from .._docker import DockerClient, PortMap, GearEnvironment
 from ..testtools import if_docker_configured, wait_for_unit_state
