+++ flocker/node/test/test_deploy.py
@@ -20,9 +20,9 @@ from .._deploy import (
     CreateVolume, WaitForVolume, HandoffVolume, SetProxies, PushVolume,
     _link_environment)
 from .._model import AttachedVolume
-from ..gear import (
-    FakeDockerClient, AlreadyExists, Unit, PortMap, GearEnvironment)
-from .._docker import DockerClient
+from .._docker import (
+    FakeDockerClient, AlreadyExists, Unit, PortMap, GearEnvironment,
+    DockerClient)
 from ...route import Proxy, make_memory_network
 from ...route._iptables import HostNetwork
 from ...volume.service import Volume
