+++ flocker/node/test/test_script.py
@@ -17,7 +17,7 @@ from ...volume.test.test_script import make_volume_options_tests
 from ..script import (
     ChangeStateOptions, ChangeStateScript,
     ReportStateScript, ReportStateOptions)
-from ..gear import FakeDockerClient, Unit
+from .._docker import FakeDockerClient, Unit
 from .._deploy import Deployer
 from .._model import Application, Deployment, DockerImage, Node, AttachedVolume
